+++ src/test/resources/features/US10_FileModule.feature
@@ -1,11 +1,11 @@
-@wipFile
+
 Feature: As a user, I should be able to delete a file/folder.
 
   Background:
     Given : user logged in successfully and on the home page
-
+  @wipFile
   Scenario: Verify users delete a file/folder
-    When the user clicks the "Files" module
+    When the user clicks on the "Files" module
     And user click action-icon from any file on the page
     And user choose the "Delete" folder option
     When the user clicks the Deleted folder sub-module on the left side
